feat: add `eol::convert_to_worktree()`.
It's the inverse of `eol::convert_to_git()` to re-add CRLF where there were LF only.
