+++ gix-filter/src/eol/mod.rs
@@ -2,6 +2,9 @@
 pub mod convert_to_git;
 pub use convert_to_git::function::convert_to_git;
 
+mod convert_to_worktree;
+pub use convert_to_worktree::convert_to_worktree;
+
 mod utils;
 
 /// The kind of end of lines to set.
